Refactor : 패키지 구조 변경, 클래스 이름 변경
- 빠뜨린 파일 추가로 구조 변경
